facturante funcionando, con la lista ya andando, no se puede probar del todo porque esta ciado el entorno de testing

+++ Omega3/Omega3/Controlador/ControlVentas.cs
@@ -286,7 +286,7 @@ namespace Omega3.Controlador
             return fecha;
         }
 
-        public static int chequearStock(int cod) {
+        public static int chequearStock(long cod) {
 
             int stock = 0;
             MySqlCommand _comando = new MySqlCommand(String.Format(
